// package/src/types.ts
export type Employee = {
  id: string;
  client_id?: string;
  emp_id: string;
  pin?: string;
  first_name: string;
  last_name: string;
  start_date: string;
  id_number?: string;
  passport?: string;
  dob: string;
  cell?: string;
  email?: string;
  job_title: string;
  department: string;
  address1?: string;
  address2?: string;
  address3?: string;
  address4?: string;
  street_number?: string;
  residency?: string;
  province?: string;
  country_of_issue?: string;
  postal_code?: string;
  tax_number?: string;
  bank_name?: string;
  portal_enabled?: 'yes' | 'no' | '';
  account_holder?: string;
  account_no?: string;
  account_type?: string;
  classification?: string;
  pay_rate: number;
  ismibco?: 'yes' | 'no' | '';
  isunion?: 'yes' | 'no' | '';
  union_name?: string;
  paye_credit?: string;
  annual_leave?: number;
  sick_leave?: number;
  family_leave?: number;
  status?: 'active' | 'offboarded';
  last_worked?: string;
  last_worked_date?: string;
  delete_reason?: string;
  image?: string; // base64 or URL
  fallbackImage?: string; // client fallback logo
};
// ...
export type LeaveType = 'annual' | 'sick' | 'family' | 'unpaid' | 'half_day';